+++ movement/movement_faces.h
@@ -75,8 +75,8 @@
 #include "rpn_calculator_face.h"
 #include "ships_bell_face.h"
 #include "habit_face.h"
-#include "planetary_face.h"
-#include "next_planetary_face.h"
+#include "planetary_time_face.h"
+#include "planetary_hours_face.h"
 // New includes go above this line.
 
 #endif // MOVEMENT_FACES_H_
